wicarproject/requirements.txt
alembic==0.9.8
amqp==2.2.2
argh==0.26.2
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.3
boto3==1.5.31
botocore==1.8.45
cachetools==2.0.1
celery==4.2.0rc3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.5.1
cryptography==2.2.2
decorator==4.2.1
docutils==0.14
dominate==2.3.1
facebook-sdk==2.0.0
flake8==3.5.0
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Celery-Helper==1.1.0
Flask-Cors==3.0.3
Flask-Markdown==0.3
Flask-Migrate==2.1.1
flask-mongoengine==0.9.5
Flask-PyMongo==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
freezegun==0.3.10
GeoAlchemy2==0.4.2
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.7.1
httplib2==0.10.3
iamport-rest-client==0.7.0
idna==2.6
infinity==1.4
intervals==0.8.1
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
log4mongo==1.6.2
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.15.0
oauth2client==4.1.2
oauthlib==2.0.6
pathtools==0.1.2
Pillow==5.0.0
psycopg2==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.6.4
pymongo==3.6.0
PyMySQL==0.8.0
python-dateutil==2.6.1
python-editor==1.0.3
python-resize-image==1.1.11
python-slugify==1.2.4
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.3
SQLAlchemy-Utils==0.33.0
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
validators==0.12.1
vine==1.1.4
visitor==0.1.3
Wand==0.4.4
Werkzeug==0.14.1
WTForms==2.1
WTForms-Alchemy==0.16.6
WTForms-Components==0.10.3
WTForms-JSON==0.3.3
WTForms-SQLAlchemy==0.1
